Proper handelling of the tracker. Skip now affects all the updates, not just the cyber-glove
